feat: bridge
view pull request [#19](https://codeberg.org/pparaxan/Quark/pulls/19) for more; with the help of AI and a week of copium, Quark can finally talk to the frontend **from** the backend and vice versa. This is not final tho :3. You can run the new example, reportsysinfo, to test this out yourself!